refactor(esphome): give both components one BLE transport (#61)
madoka_vam.cpp carried a copy of madoka.cpp's whole transport layer. Seven
functions were identical character for character -- loop, query_,
split_payload_, prepare_message_, process_incoming_chunk_, set_eye_brightness
and reset_filter -- and four more differed only in a log label or a command id.
The predictable outcome was the next transport fix landing in one copy while
the other drifted, with CI unable to notice because it only compiles these
components, it never runs them.

A new madoka_base component holds everything that does not depend on what is
wired behind the BRC1H: the GATT plumbing, the 20-byte chunking, the poll loop,
and the three features every unit exposes (clean filter, firmware version, eye
brightness). Madoka and MadokaVam derive from it and supply four things each --
their log tag, their display name, how to decode a message, and which extra
state to poll. Two more hooks cover the rest: on_setup_ for the VAM's custom
presets, on_disconnect_ for the thermostat's setpoint reset.

The poll order is preserved exactly. update() sends the two shared queries,
calls query_appliance_state_() for the setpoint/fan pair or the ventilation
read, then the four trailing shared ones -- same commands, same delays, same
sequence as before on both components.

1398 lines become 1225, and the ~380 lines of transport exist once.

## This changes user YAML

ESPHome only copies the external components listed in `components:`, and
AUTO_LOAD cannot reach one that is not listed, so `madoka_base` has to be named
alongside `madoka` or `madoka_vam`. Every documented config is updated and the
README carries an upgrade note. A node keeps running its current firmware until
it is recompiled, and a forgotten entry fails the build with
`Component not found: madoka_base` rather than misbehaving.

## Not done here

The blocking `esphome::delay(t_d)` in query_ is left as it was, against the
issue's suggestion to remove it in the same pass. It now exists once, which was
the precondition, but removing it means pacing writes from loop() and changing
when responses arrive -- a behaviour change on a path neither CI nor any
hardware here can exercise, since there is no VAM to test against. The comment
in query_ says so.

Validated with `esphome config` on all three test configurations against
ESPHome 2026.8.2. The firmware compile is CI's.
