+++ esphome/components/madoka/madoka.h
@@ -2,26 +2,12 @@
 
 #include <cmath>
 #include <vector>
-#include <queue>
-#include <map>
 #include <string>
 
-#include "esphome/core/component.h"
-#include "esphome/components/binary_sensor/binary_sensor.h"
-#include "esphome/components/ble_client/ble_client.h"
-#include "esphome/components/button/button.h"
-#include "esphome/components/esp32_ble_tracker/esp32_ble_tracker.h"
-#include "esphome/components/climate/climate.h"
-#include "esphome/components/number/number.h"
+#include "esphome/components/madoka_base/madoka_base.h"
 #include "esphome/components/sensor/sensor.h"
-#include "esphome/components/text_sensor/text_sensor.h"
 
 #ifdef USE_ESP32
-
-#include <esp_gattc_api.h>
-
-static const uint8_t MAX_CHUNK_SIZE = 20;
-static const uint8_t BLE_SEND_MAX_RETRIES = 5;
 
 namespace esphome {
 namespace madoka {
@@ -53,22 +39,10 @@ struct SensorReading {
   uint8_t outdoor;
 };
 
-struct Status {
-  bool status;
-  uint8_t mode;
-};
-
 namespace espbt = esphome::esp32_ble_tracker;
 
-static const espbt::ESPBTUUID MADOKA_SERVICE_UUID = espbt::ESPBTUUID::from_raw("2141e110-213a-11e6-b67b-9e71128cae77");
-static const espbt::ESPBTUUID NOTIFY_CHARACTERISTIC_UUID =
-    espbt::ESPBTUUID::from_raw("2141e111-213a-11e6-b67b-9e71128cae77");
-static const espbt::ESPBTUUID WWR_CHARACTERISTIC_UUID =
-    espbt::ESPBTUUID::from_raw("2141e112-213a-11e6-b67b-9e71128cae77");
-
-class Madoka : public climate::Climate, public esphome::ble_client::BLEClientNode, public PollingComponent {
+class Madoka : public madoka_base::MadokaBase {
  protected:
-  bool should_update_ = false;
   // Whether the entity advertises two setpoints (BRC1H "range" mode) or a
   // single one. ESPHome traits are static per API connection, so this is a
   // YAML option rather than something derived from the device at runtime.
@@ -84,48 +58,21 @@ class Madoka : public climate::Climate, public esphome::ble_client::BLEClientNod
   // has to carry the gap. Defaults to 0 so a controller that does not report
   // the argument keeps the equal-pair behaviour.
   uint8_t min_differential_ = 0;
-  std::queue<std::vector<uint8_t>> received_chunks_ = {};
-  std::map<uint8_t, std::vector<uint8_t>> pending_chunks_ = {};
-  uint16_t notify_handle_;
-  uint16_t wwr_handle_;
-  SemaphoreHandle_t receive_semaphore_ = nullptr;
-  Status cur_status_;
   sensor::Sensor *outdoor_temperature_sensor_{nullptr};
-  binary_sensor::BinarySensor *clean_filter_binary_sensor_{nullptr};
-  text_sensor::TextSensor *firmware_version_text_sensor_{nullptr};
-  number::Number *eye_brightness_number_{nullptr};
-  button::Button *reset_filter_button_{nullptr};
 
-  std::vector<std::vector<uint8_t>> split_payload_(std::vector<uint8_t> msg);
-  std::vector<uint8_t> prepare_message_(uint16_t cmd, std::vector<uint8_t> args);
-  void query_(uint16_t cmd, std::vector<uint8_t> args, int t_d);
-  void parse_cb_(std::vector<uint8_t> msg);
-  void process_incoming_chunk_(std::vector<uint8_t> chk);
+  const char *tag_() const override;
+  const char *label_() const override { return "Daikin Madoka Climate Controller"; }
+  void parse_cb_(std::vector<uint8_t> msg) override;
+  void query_appliance_state_() override;
+  void on_disconnect_() override;
+
   void apply_setpoints_();
 
   void control(const climate::ClimateCall &call) override;
 
  public:
-  void setup() override;
-  void loop() override;
-  void update() override;
-  void gattc_event_handler(esp_gattc_cb_event_t event, esp_gatt_if_t gattc_if,
-                           esp_ble_gattc_cb_param_t *param) override;
-  void gap_event_handler(esp_gap_ble_cb_event_t event, esp_ble_gap_cb_param_t *param) override;
-  void dump_config() override;
   void set_outdoor_temperature_sensor(sensor::Sensor *sensor) { this->outdoor_temperature_sensor_ = sensor; }
-  void set_clean_filter_binary_sensor(binary_sensor::BinarySensor *sensor) {
-    this->clean_filter_binary_sensor_ = sensor;
-  }
-  void set_firmware_version_text_sensor(text_sensor::TextSensor *sensor) {
-    this->firmware_version_text_sensor_ = sensor;
-  }
   void set_dual_setpoint(bool dual_setpoint) { this->dual_setpoint_ = dual_setpoint; }
-  void set_eye_brightness_number(number::Number *number) { this->eye_brightness_number_ = number; }
-  void set_reset_filter_button(button::Button *button) { this->reset_filter_button_ = button; }
-  void set_eye_brightness(uint8_t level);
-  void reset_filter();
-  float get_setup_priority() const override { return setup_priority::DATA; }
   climate::ClimateTraits traits() override {
     auto traits = climate::ClimateTraits();
     traits.set_supported_modes({
