+++ esphome/components/madoka_vam/climate.py
@@ -12,17 +12,23 @@ from esphome.const import CONF_ID, DEVICE_CLASS_PROBLEM
 
 CODEOWNERS = ["@Frank802"]
 DEPENDENCIES = ["ble_client"]
-AUTO_LOAD = ["binary_sensor", "button", "number", "text_sensor"]
+AUTO_LOAD = ["binary_sensor", "button", "madoka_base", "number", "text_sensor"]
 
 CONF_CLEAN_FILTER = "clean_filter"
 CONF_FIRMWARE_VERSION = "firmware_version"
 CONF_EYE_BRIGHTNESS = "eye_brightness"
 CONF_RESET_FILTER = "reset_filter"
 
-madoka_vam_ns = cg.esphome_ns.namespace("madoka_vam")
-MadokaVam = madoka_vam_ns.class_(
-    "MadokaVam", climate.Climate, ble_client.BLEClientNode, cg.PollingComponent
+# Declared here rather than imported from the madoka_base component: an
+# external component that is not itself listed in `external_components:`
+# cannot be imported by name. This is type metadata for codegen only --
+# the shared implementation lives in madoka_base/madoka_base.h.
+madoka_base_ns = cg.esphome_ns.namespace("madoka_base")
+MadokaBase = madoka_base_ns.class_(
+    "MadokaBase", climate.Climate, ble_client.BLEClientNode, cg.PollingComponent
 )
+madoka_vam_ns = cg.esphome_ns.namespace("madoka_vam")
+MadokaVam = madoka_vam_ns.class_("MadokaVam", MadokaBase)
 MadokaEyeBrightnessNumber = madoka_vam_ns.class_(
     "MadokaEyeBrightnessNumber", number.Number, cg.Parented.template(MadokaVam)
 )
