+++ esphome/components/madoka_vam/madoka_vam.h
@@ -1,28 +1,14 @@
 #pragma once
 
 #include <vector>
-#include <queue>
-#include <map>
 #include <string>
 
-#include "esphome/core/component.h"
-#include "esphome/components/binary_sensor/binary_sensor.h"
-#include "esphome/components/ble_client/ble_client.h"
-#include "esphome/components/button/button.h"
-#include "esphome/components/esp32_ble_tracker/esp32_ble_tracker.h"
-#include "esphome/components/climate/climate.h"
-#include "esphome/components/number/number.h"
-#include "esphome/components/text_sensor/text_sensor.h"
+#include "esphome/components/madoka_base/madoka_base.h"
 
 #ifdef USE_ESP32
 
-#include <esp_gattc_api.h>
-
 namespace esphome {
 namespace madoka_vam {
-
-static const uint8_t MAX_CHUNK_SIZE = 20;
-static const uint8_t BLE_SEND_MAX_RETRIES = 5;
 
 class MadokaVam;
 
@@ -36,64 +22,29 @@ class MadokaResetFilterButton : public button::Button, public Parented<MadokaVam
   void press_action() override;
 };
 
-struct Status {
-  bool status;
-  uint8_t mode;
-};
-
 namespace espbt = esphome::esp32_ble_tracker;
 
-static const espbt::ESPBTUUID MADOKA_SERVICE_UUID = espbt::ESPBTUUID::from_raw("2141e110-213a-11e6-b67b-9e71128cae77");
-static const espbt::ESPBTUUID NOTIFY_CHARACTERISTIC_UUID =
-    espbt::ESPBTUUID::from_raw("2141e111-213a-11e6-b67b-9e71128cae77");
-static const espbt::ESPBTUUID WWR_CHARACTERISTIC_UUID =
-    espbt::ESPBTUUID::from_raw("2141e112-213a-11e6-b67b-9e71128cae77");
-
-class MadokaVam : public climate::Climate, public esphome::ble_client::BLEClientNode, public PollingComponent {
+/// A VAM (Ventilation Air Management / heat-recovery unit) behind a BRC1H.
+/// It only ventilates: no setpoint, no outdoor probe, and the air routing is
+/// exposed as a custom preset rather than an HVAC mode.
+class MadokaVam : public madoka_base::MadokaBase {
  protected:
-  bool should_update_ = false;
-  std::queue<std::vector<uint8_t>> received_chunks_ = {};
-  std::map<uint8_t, std::vector<uint8_t>> pending_chunks_ = {};
-  uint16_t notify_handle_;
-  uint16_t wwr_handle_;
-  SemaphoreHandle_t receive_semaphore_ = nullptr;
-  Status cur_status_;
   bool dump_raw_ = false;
-  binary_sensor::BinarySensor *clean_filter_binary_sensor_{nullptr};
-  text_sensor::TextSensor *firmware_version_text_sensor_{nullptr};
-  number::Number *eye_brightness_number_{nullptr};
 
-  std::vector<std::vector<uint8_t>> split_payload_(std::vector<uint8_t> msg);
-  std::vector<uint8_t> prepare_message_(uint16_t cmd, std::vector<uint8_t> args);
-  void query_(uint16_t cmd, std::vector<uint8_t> args, int t_d);
-  void parse_cb_(std::vector<uint8_t> msg);
-  void process_incoming_chunk_(std::vector<uint8_t> chk);
+  const char *tag_() const override;
+  const char *label_() const override { return "Daikin Madoka VAM Climate Controller"; }
+  void parse_cb_(std::vector<uint8_t> msg) override;
+  void query_appliance_state_() override;
+  void on_setup_() override;
 
   void control(const climate::ClimateCall &call) override;
 
  public:
-  void setup() override;
-  void loop() override;
-  void update() override;
-  void gattc_event_handler(esp_gattc_cb_event_t event, esp_gatt_if_t gattc_if,
-                           esp_ble_gattc_cb_param_t *param) override;
-  void gap_event_handler(esp_gap_ble_cb_event_t event, esp_ble_gap_cb_param_t *param) override;
-  void dump_config() override;
-  void set_clean_filter_binary_sensor(binary_sensor::BinarySensor *sensor) {
-    this->clean_filter_binary_sensor_ = sensor;
-  }
-  void set_firmware_version_text_sensor(text_sensor::TextSensor *sensor) {
-    this->firmware_version_text_sensor_ = sensor;
-  }
-  void set_eye_brightness_number(number::Number *number) { this->eye_brightness_number_ = number; }
-  void set_eye_brightness(uint8_t level);
-  void reset_filter();
   void set_dump_raw(bool dump_raw) { this->dump_raw_ = dump_raw; }
-  // Envoi d'une commande brute (function id + arguments) : utile pour sonder des
-  // fonctions non documentées du VAM depuis une lambda ESPHome. Les réponses
-  // arrivent dans les logs (activer dump_raw pour l'affichage hexadécimal).
+  // Send a raw command (function id plus arguments): useful for probing the
+  // VAM's undocumented functions from an ESPHome lambda. Responses land in the
+  // logs; enable dump_raw for the hex dump.
   void send_raw_command(uint16_t cmd, std::vector<uint8_t> args) { this->query_(cmd, std::move(args), 200); }
-  float get_setup_priority() const override { return setup_priority::DATA; }
   climate::ClimateTraits traits() override {
     auto traits = climate::ClimateTraits();
     traits.set_supported_modes({
